backend/pyproject.toml
[project]
name = "stocksage-api"
version = "0.1.0"
description = "FastAPI backend for StockSage-AI stock trading simulator"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyrebase4 (>=4.8.0,<5.0.0)",
    "setuptools (>=75.8.2,<76.0.0)",
    "firebase-admin (>=6.7.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
